requirements.txt
digimat.saia
python-dotenv
netifaces

protobuf~=6.31.1
grpcio
